feat(switch): migrate to the new component model
