+++ packages/admin-ui/src/switch/index.ts
@@ -1,3 +1,1 @@
 export * from './switch'
-export * from './switch-button'
-export * from './types'
